+++ node_blog_backend/controllers/repliesController.js
@@ -21,7 +21,13 @@ const getReply = async (req, res) => {
 
 const createReply = async (req, res) => {
   try {
-      // Crear la nueva respuesta
+      
+      // throw new Error("Simulated error in getUserPosts");
+      const comment = await Comment.findById(req.params.id);
+      if (!comment) {
+          return res.status(404).json({ error: 'Error', msg: "Comment not found" });
+      }
+
       const reply = new Replies({
           reply: req.body.reply,
           commentId: req.params.id,
@@ -30,16 +36,8 @@ const createReply = async (req, res) => {
           dateReply: new Date()
       });
 
-      // Guardar la nueva respuesta
       const newReply = await reply.save();
 
-      // Buscar el comentario correspondiente
-      const comment = await Comment.findById(req.params.id);
-      if (!comment) {
-          return res.status(404).json({ message: "Comment not found" });
-      }
-
-      // Añadir la nueva respuesta al array de respuestas del comentario
       comment.replies.push(newReply._id);
       await comment.save();
 
@@ -58,14 +56,24 @@ const createReply = async (req, res) => {
         })
         res.json(comment2);
   } catch (error) {
-      res.status(400).json({ message: error.message });
+      res.status(400).json({  error: 'Error', msg: error.message });
   }
 };
 
 
 const updateReply = async (req, res) => {
   try {
+    // throw new Error("Simulated error in getUserPosts");
     const reply = await Replies.findById(req.params.id);
+
+    if(!reply) {
+      return res.status(404).json({ error: 'Error', msg: "Reply not found" });
+    }
+
+    if (reply.userID.toString() !== req.query.user) {
+        return res.status(401).json({ error: 'Error', msg: "Unauthorized" });
+    }
+
     if (req.body.reply) {
       reply.reply = req.body.reply;
     }
@@ -85,12 +93,24 @@ const updateReply = async (req, res) => {
         })
         res.json(comment2);
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    res.status(500).json({ error: 'Error', msg: error.message });
   }
 };
 
 const deleteReply = async (req, res) => {
   try {
+
+    // throw new Error("Simulated error in getUserPosts");
+    const reply = await Replies.findById(req.params.id);
+
+    if(!reply) {
+      return res.status(404).json({ error: 'Error', msg: "Reply not found" });
+    }
+
+    if (reply.userID.toString() !== req.query.user) {
+        return res.status(401).json({ error: 'Error', msg: "Unauthorized" });
+    }
+
     await Replies.findByIdAndDelete(req.params.id);
     const comment2 = await Comment.findById(req.body.commentID)
         .select('comment dateComment ')  
@@ -107,7 +127,7 @@ const deleteReply = async (req, res) => {
         })
         res.json(comment2);
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    res.status(500).json({  error: 'Error', msg: error.message});
   }
 };
